Hooked up missing VRIME keyboard bindings. Show VRIME keyboard after 3 seconds. Pulled over some other bits from the VRIME SteamVR 2 sample, that may not be strictly necessary.

+++ Source/FirefoxRealityProUnity/Assets/VRIME2/Scripts/InputField/VRIME_CursorTouchPad.cs
@@ -55,13 +55,14 @@ namespace VRIME2
 #endif
 
 		private void CheckTouchPadPress() {
-#if steamvr_v2
-			int leftIndex = VRIME_Manager.Ins.userControllerLeft.GetDeviceIndex();
-			int rightIndex = VRIME_Manager.Ins.userControllerRight.GetDeviceIndex();
-#else			
+#if steamvr_v2			
+			int leftIndex = (int)VRIME_Manager.Ins.userControllerLeft.GetDeviceIndex();
+			int rightIndex = (int)VRIME_Manager.Ins.userControllerRight.GetDeviceIndex();
+#else
 			int leftIndex = (int)VRIME_Manager.Ins.userControllerLeft.index;
 			int rightIndex = (int)VRIME_Manager.Ins.userControllerRight.index;
 #endif			
+
 
 			bool isLeftPressTouched = false;
 			bool isRightPressTouched = false;
